requirements.txt
# ==========================================================
# 📦 DEPENDÊNCIAS PRINCIPAIS DO PROJETO
# Grupo SBF - Analytics Engineer Case
# Autora: Camila Macedo
# ==========================================================

# --- Manipulação e processamento de dados ---
pandas==2.2.3
numpy==1.26.4

# --- Banco de dados (PostgreSQL via SQLAlchemy) ---
SQLAlchemy==1.4.52         # ✅ compatível com Airflow 2.10.x
psycopg2-binary==2.9.9

# --- Gerenciamento de variáveis de ambiente ---
python-dotenv==1.0.1

# --- Orquestração local (Airflow) ---
apache-airflow==2.10.2

# --- Logs e monitoramento (opcional / recomendado) ---
rich==13.9.2

# --- Compatibilidade e segurança ---
requests==2.32.3

# --- Utilitários de data/hora ---
python-dateutil==2.9.0.post0

# ==========================================================
# 🧠 NOTAS
# ----------------------------------------------------------
# SQLAlchemy 1.4.52 → versão estável suportada pelo Airflow
# ==========================================================
